requirements.txt
aioredis==1.3.1
asgiref==3.2.7
async-timeout==2.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
bitmex==0.2.2
bitmex-ws==0.4.0
bravado==10.6.2
bravado-core==5.17.0
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==3.0.7
django-crispy-forms==1.9.1
djangorestframework==3.11.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
jsonpointer==2.0
jsonref==0.2
jsonschema==3.2.0
monotonic==1.5
msgpack==0.6.2
msgpack-python==0.5.6
numpy==1.18.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
rfc3987==1.3.8
service-identity==18.1.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
strict-rfc3339==0.7
swagger-spec-validator==2.6.0
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
webcolors==1.11.1
websocket-client==0.53.0
zipp==3.1.0
zope.interface==5.1.0
